ToDoApp
A to do app project was made using MVC. First of all, the login
page and tables for the items to be added to the to-do list were created
in SQL, then the page that should be opened after the user logged in was
created. A separate title has been created for completed items at the
bottom of the page, and a button has been created to add a new item.
For existing items, the edit, delete and details buttons have been created
next to them. A checkbox was placed to the left of the items. When
clicking on the checbox, the item is set to go to the completed items
section under the completed items heading, and when clicking on the
checbox again, it will return to the incomplete state. At the same time,
date, detail and priority information can be seen next to the titles of the
items on the main page.